envelope-api/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
reportlab
# Módulo IA Compras
pymongo==4.7.0
motor==3.4.0
httpx==0.27.0
beautifulsoup4==4.12.3
langextract>=0.1.0
google-genai>=1.0.0
